+++ app/src/androidTest/java/ch/epfl/sweng/SDP/Matchmaking/MatchmakerTest.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.SDP.matchmaking;
+package ch.epfl.sweng.SDP.Matchmaking;
 
 import static org.junit.Assert.assertTrue;
