+++ app/src/main/java/ch/epfl/sweng/SDP/Matchmaking/CurrentUser.java
@@ -1,4 +1,4 @@
-package ch.epfl.sweng.SDP.matchmaking;
+package ch.epfl.sweng.SDP.Matchmaking;
 
 import com.google.firebase.auth.FirebaseAuth;
